Copied several relevant Django apps from my JBA repository.
Django tutorial polls app: Copied because that's what we are currently using for practice in the JBM course.
user_manager app: As reference and starter for dealing with user related stuff.
site_index app: To deal with the multiple apps.
django_testing app: To copy the relevant settings for the site_index app.